splises files
